Create index.html
Started working on Lab 1 configured the html structure added title tag with the title of the class. Added h1 tag again with the name of the class. Added 1 h3 tags for the semester. 2 h2 tags for the professor and time and place heading. And 3 p tags for my name, professor's name, and location and time of the class.